User record refactor & ArchTest
