+++ src/main/java/org/example/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package org.example.repository;
 
-import org.example.User;
+import org.example.users.User;
 
 import java.security.NoSuchAlgorithmException;
