+++ src/main/java/org/example/users/User.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.users;
 
 public record User(String username, String password) {
 }
