--- src/test/java/org/example/AppIT.java
@@ -1,12 +1,0 @@
-package org.example;
-
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class AppIT {
-    @Test
-    void itTest() {
-        assertThat(false).isFalse();
-    }
-}
