--- src/test/java/org/example/AppTest.java
@@ -1,12 +1,0 @@
-package org.example;
-
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class AppTest {
-    @Test
-    void test() {
-        assertThat(true).isTrue();
-    }
-}
